* Rewrite scheduler by scala * Simplify code * Upgrade libraries

+++ mycollab-core/src/main/java/com/esofthead/mycollab/core/utils/ImageUtil.java
@@ -133,7 +133,7 @@ public class ImageUtil {
 			int height = (int) (imgHeight / scale);
 
 			BufferedImage rescaledImage = Scalr.resize(source, Method.QUALITY,
-					Mode.AUTOMATIC, idealWidth, (int) height);
+					Mode.AUTOMATIC, idealWidth, height);
 			if (height > 400) {
 				rescaledImage = rescaledImage.getSubimage(0, 0,
 						Math.min(256, rescaledImage.getWidth()), 400);
